Moved functions to pytorch_utils.py
